Update changes module settings
