+++ src/app/pages/planner/pagination-table/pagination-table.component.ts
@@ -1,0 +1,229 @@
+import { Component, ViewChild } from '@angular/core';
+import { MatPaginator, MatPaginatorModule } from '@angular/material/paginator';
+import { MatTableDataSource, MatTableModule } from '@angular/material/table';
+import { BreakpointObserver } from '@angular/cdk/layout';
+import { CommonModule } from '@angular/common';
+import { MatCardModule } from '@angular/material/card';
+
+const PRODUCT_DATA: Element[] = [
+  {
+    id: 1,
+    imagePath: 'assets/images/profile/user-1.jpg',
+    uname: 'Sunil Joshi',
+    position: 'Web Designer',
+    productName: 'Elite Admin',
+    budget: 3.9,
+    priority: 'low',
+  },
+  {
+    id: 2,
+    imagePath: 'assets/images/profile/user-2.jpg',
+    uname: 'Andrew McDownland',
+    position: 'Project Manager',
+    productName: 'Real Homes Theme',
+    budget: 24.5,
+    priority: 'medium',
+  },
+  {
+    id: 3,
+    imagePath: 'assets/images/profile/user-3.jpg',
+    uname: 'Christopher Jamil',
+    position: 'Project Manager',
+    productName: 'MedicalPro Theme',
+    budget: 12.8,
+    priority: 'high',
+  },
+  {
+    id: 4,
+    imagePath: 'assets/images/profile/user-4.jpg',
+    uname: 'Nirav Joshi',
+    position: 'Frontend Engineer',
+    productName: 'Hosting Press HTML',
+    budget: 2.4,
+    priority: 'critical',
+  },
+  {
+    id: 1,
+    imagePath: 'assets/images/profile/user-1.jpg',
+    uname: 'Sunil Joshi',
+    position: 'Web Designer',
+    productName: 'Elite Admin',
+    budget: 3.9,
+    priority: 'low',
+  },
+  {
+    id: 2,
+    imagePath: 'assets/images/profile/user-2.jpg',
+    uname: 'Andrew McDownland',
+    position: 'Project Manager',
+    productName: 'Real Homes Theme',
+    budget: 24.5,
+    priority: 'medium',
+  },
+  {
+    id: 3,
+    imagePath: 'assets/images/profile/user-3.jpg',
+    uname: 'Christopher Jamil',
+    position: 'Project Manager',
+    productName: 'MedicalPro Theme',
+    budget: 12.8,
+    priority: 'high',
+  },
+  {
+    id: 4,
+    imagePath: 'assets/images/profile/user-4.jpg',
+    uname: 'Nirav Joshi',
+    position: 'Frontend Engineer',
+    productName: 'Hosting Press HTML',
+    budget: 2.4,
+    priority: 'critical',
+  },
+  {
+    id: 1,
+    imagePath: 'assets/images/profile/user-1.jpg',
+    uname: 'Sunil Joshi',
+    position: 'Web Designer',
+    productName: 'Elite Admin',
+    budget: 3.9,
+    priority: 'low',
+  },
+  {
+    id: 2,
+    imagePath: 'assets/images/profile/user-2.jpg',
+    uname: 'Andrew McDownland',
+    position: 'Project Manager',
+    productName: 'Real Homes Theme',
+    budget: 24.5,
+    priority: 'medium',
+  },
+  {
+    id: 3,
+    imagePath: 'assets/images/profile/user-3.jpg',
+    uname: 'Christopher Jamil',
+    position: 'Project Manager',
+    productName: 'MedicalPro Theme',
+    budget: 12.8,
+    priority: 'high',
+  },
+  {
+    id: 4,
+    imagePath: 'assets/images/profile/user-4.jpg',
+    uname: 'Nirav Joshi',
+    position: 'Frontend Engineer',
+    productName: 'Hosting Press HTML',
+    budget: 2.4,
+    priority: 'critical',
+  },
+  {
+    id: 1,
+    imagePath: 'assets/images/profile/user-1.jpg',
+    uname: 'Sunil Joshi',
+    position: 'Web Designer',
+    productName: 'Elite Admin',
+    budget: 3.9,
+    priority: 'low',
+  },
+  {
+    id: 2,
+    imagePath: 'assets/images/profile/user-2.jpg',
+    uname: 'Andrew McDownland',
+    position: 'Project Manager',
+    productName: 'Real Homes Theme',
+    budget: 24.5,
+    priority: 'medium',
+  },
+  {
+    id: 3,
+    imagePath: 'assets/images/profile/user-3.jpg',
+    uname: 'Christopher Jamil',
+    position: 'Project Manager',
+    productName: 'MedicalPro Theme',
+    budget: 12.8,
+    priority: 'high',
+  },
+  {
+    id: 4,
+    imagePath: 'assets/images/profile/user-4.jpg',
+    uname: 'Nirav Joshi',
+    position: 'Frontend Engineer',
+    productName: 'Hosting Press HTML',
+    budget: 2.4,
+    priority: 'critical',
+  },
+  {
+    id: 1,
+    imagePath: 'assets/images/profile/user-1.jpg',
+    uname: 'Sunil Joshi',
+    position: 'Web Designer',
+    productName: 'Elite Admin',
+    budget: 3.9,
+    priority: 'low',
+  },
+  {
+    id: 2,
+    imagePath: 'assets/images/profile/user-2.jpg',
+    uname: 'Andrew McDownland',
+    position: 'Project Manager',
+    productName: 'Real Homes Theme',
+    budget: 24.5,
+    priority: 'medium',
+  },
+  {
+    id: 3,
+    imagePath: 'assets/images/profile/user-3.jpg',
+    uname: 'Christopher Jamil',
+    position: 'Project Manager',
+    productName: 'MedicalPro Theme',
+    budget: 12.8,
+    priority: 'high',
+  },
+  {
+    id: 4,
+    imagePath: 'assets/images/profile/user-4.jpg',
+    uname: 'Nirav Joshi',
+    position: 'Frontend Engineer',
+    productName: 'Hosting Press HTML',
+    budget: 2.4,
+    priority: 'critical',
+  },
+];
+
+@Component({
+  selector: 'app-pagination-table',
+  standalone: true,
+  imports: [MatCardModule, MatTableModule, CommonModule, MatPaginatorModule],
+  templateUrl: './pagination-table.component.html',
+})
+export class AppPaginationTableComponent {
+  displayedColumns = ['assigned', 'name', 'priority', 'budget'];
+  dataSource = new MatTableDataSource<Element>(PRODUCT_DATA);
+
+  constructor(breakpointObserver: BreakpointObserver) {
+    breakpointObserver.observe(['(max-width: 600px)']).subscribe((result) => {
+      this.displayedColumns = result.matches
+        ? ['assigned', 'name', 'priority', 'budget']
+        : ['assigned', 'name', 'priority', 'budget'];
+    });
+  }
+
+  @ViewChild(MatPaginator, { static: true }) paginator: MatPaginator =
+    Object.create(null);
+
+  /**
+   * Set the paginator after the view init since this component will
+   * be able to query its view for the initialized paginator.
+   */
+  ngAfterViewInit(): void {
+    this.dataSource.paginator = this.paginator;
+  }
+}
+
+export interface Element {
+  id: number;
+  imagePath: string;
+  uname: string;
+  position: string;
+  productName: string;
+  budget: number;
+  priority: string;
+}
