+++ src/app/pages/settings/settings.routes.ts
@@ -1,0 +1,17 @@
+import { Routes } from "@angular/router";
+import { UsersComponent } from "./users/users.component";
+
+export const SettingsRoutes: Routes = [
+    {
+      path: '',
+      children: [
+        {
+          path: 'users',
+          component:  UsersComponent,
+          data: {
+            title: 'Usuarios',
+          },
+        },
+      ],
+    },
+  ];
